pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "routine-calculator"
version = "1.0.0"
description = "A console application to track where you spend your time"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "python-dotenv==1.0.0",
    "PyYAML==6.0.3",
    "simpleeval==0.9.13",
]

[project.scripts]
routine-calculator = "routine_calculator.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["routine_calculator*"]
